backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-multipart==0.0.12
httpx==0.27.2
pytest==8.3.3
openai>=1.50.0
python-dotenv>=1.0.0
